Extract helper functions from tif loader
